EclecticIQ/requirements.txt
###### Requirements without Version Specifiers ######
pandas
robotframework
robotframework-selenium2library
